// animacionweb3d/src/typings.d.ts
export type DroneGLTF = GLTF & {
    nodes : {
        Cube004 : THREE.Mesh;
        Cube004_DarkSilver_0 : THREE.Mesh;
        Cube004_Dustygrey_0 : THREE.Mesh;
        Cube004_Glowglowblue_0 : THREE.Mesh;
        Cube004_Glowred_0 : THREE.Mesh;
        Cube004_Gold_0 : THREE.Mesh;
        Cube004_Silver_0 : THREE.Mesh;
        Cube004_White_0 : THREE.Mesh;
        RootNode : THREE.Mesh;
        Sketchfab_Scene : THREE.Mesh;
        Sketchfab_model : THREE.Mesh;
    };
    materials: {
        DarkSilver : THREE.MeshStandardMaterial;
        Dustygrey : THREE.MeshStandardMaterial;
        Glowglowblue : THREE.MeshStandardMaterial;
        Glowred : THREE.MeshStandardMaterial;
        Gold : THREE.MeshStandardMaterial;
        Silver : THREE.MeshStandardMaterial;
        White : THREE.MeshStandardMaterial;
    }
}